move rendezvous status structs to nexus-types
